hdf, log

+++ src/log.py
@@ -5,14 +5,31 @@ import platform
 import sys
 
 
-def start_logging(level):
-    log_format = log.Formatter('%(asctime)s, %(levelname)s, %(message)s')
+def start_logging(level, filename, remove_old=False):
+    """
+    Starts logger for crowd dynamics simulation.
+
+    :param level:
+    :param filename:
+    :return:
+    """
+    # Create filename.log file
+    ext = ".log"
+    filename, _ = os.path.splitext(filename)
+    filename += ext
+
+    # Remove old log file
+    if remove_old and os.path.exists(filename):
+        os.remove(filename)  # Remove old log file
+
+    # Format of the log output
+    log_format = log.Formatter('%(asctime)s, '
+                               '%(levelname)s, '
+                               '%(funcName)s, '
+                               '%(message)s')
     logger = log.getLogger()
     logger.setLevel(level)
 
-    filename = "run.log"
-    if os.path.exists(filename):
-        os.remove(filename)  # Remove old log file
     file_handler = logging.handlers.RotatingFileHandler(
         filename, maxBytes=(10240 * 5), backupCount=2
     )
